apply_worker/requirements.txt
boto3
python-dotenv
gspread
google-auth
playwright
openai 